+++ weblate/trans/models/component.py
@@ -93,7 +93,6 @@ from weblate.trans.validators import (
 from weblate.utils import messages
 from weblate.utils.celery import get_task_progress
 from weblate.utils.colors import ColorChoices
-from weblate.utils.db import using_postgresql
 from weblate.utils.decorators import disable_for_loaddata
 from weblate.utils.errors import report_error
 from weblate.utils.fields import EmailField
@@ -1339,7 +1338,12 @@ class Component(
         self.updated_sources[source.id] = source
         return change
 
-    def bulk_create_sources(self, attributes_list: list[UnitAttributesDict]) -> None:
+    def bulk_create_sources(
+        self,
+        attributes_list: list[UnitAttributesDict],
+        *,
+        create_unit_change_action: ActionEvents = ActionEvents.NEW_UNIT_REPO,
+    ) -> None:
         """Ensure that all sources are stored in the database."""
         # Can not bulk create with getting primary key, resort to one by one creation
         if not connection.features.can_return_rows_from_bulk_insert:
@@ -1392,7 +1396,12 @@ class Component(
             self._sources[unit.id_hash] = unit
 
             # Postprocess and create change
-            changes.append(self._process_new_source(unit, save=False))
+            change = self._process_new_source(unit, save=False)
+            if create_unit_change_action == ActionEvents.NEW_UNIT_UPLOAD:
+                change.action = ActionEvents.NEW_SOURCE_UPLOAD
+            elif create_unit_change_action == ActionEvents.NEW_UNIT_REPO:
+                change.action = ActionEvents.NEW_SOURCE_REPO
+            changes.append(change)
 
         # Update source unit in the database
         Unit.objects.bulk_update(units, fields=["source_unit"])
@@ -3778,9 +3787,7 @@ class Component(
         queryset = PendingUnitChange.objects.for_component(
             self, apply_filters=apply_filters
         )
-        if using_postgresql():
-            return queryset.distinct("unit_id").count()
-        return queryset.values("unit_id").distinct().count()
+        return queryset.distinct("unit_id").count()
 
     @property
     def count_repo_missing(self):