+++ weblate/trans/models/translation.py
@@ -47,7 +47,6 @@ from weblate.trans.signals import component_post_update, vcs_pre_commit
 from weblate.trans.util import is_plural, join_plural, split_plural
 from weblate.trans.validators import validate_check_flags
 from weblate.utils import messages
-from weblate.utils.db import using_postgresql
 from weblate.utils.errors import report_error
 from weblate.utils.html import format_html_join_comma
 from weblate.utils.render import render_template
@@ -572,7 +571,8 @@ class Translation(
                                 newunit.unit_attributes
                                 for newunit in updated.values()
                                 if newunit.unit_attributes is not None
-                            ]
+                            ],
+                            create_unit_change_action=self.create_unit_change_action,
                         )
 
                 # Create/update translations
@@ -907,9 +907,7 @@ class Translation(
     def count_pending_units(self):
         """Return count of units with pending changes."""
         qs = PendingUnitChange.objects.for_translation(self, apply_filters=True)
-        if using_postgresql():
-            return qs.distinct("unit_id").count()
-        return qs.values("unit_id").distinct().count()
+        return qs.distinct("unit_id").count()
 
     def needs_commit(self):
         """Check whether there are some not committed changes."""