+++ weblate/trans/tests/test_middleware.py
@@ -10,7 +10,6 @@ from django.contrib.messages import get_messages
 from django.urls import reverse
 
 from weblate.trans.tests.test_views import FixtureTestCase
-from weblate.utils.db import using_postgresql
 
 
 class MiddlewareTestCase(FixtureTestCase):
@@ -22,14 +21,11 @@ class MiddlewareTestCase(FixtureTestCase):
         self.assertEqual(response.status_code, 404)
 
     def test_project_redirect(self) -> None:
-        # Different casing should redirect, MySQL always does case insensitive lookups
-        if using_postgresql():
-            response = self.client.get(
-                reverse("show", kwargs={"path": [self.project.slug.upper()]})
-            )
-            self.assertRedirects(
-                response, self.project.get_absolute_url(), status_code=301
-            )
+        # Different casing should redirect
+        response = self.client.get(
+            reverse("show", kwargs={"path": [self.project.slug.upper()]})
+        )
+        self.assertRedirects(response, self.project.get_absolute_url(), status_code=301)
 
     def test_component_redirect(self) -> None:
         # Non existing fails with 404
@@ -37,15 +33,14 @@ class MiddlewareTestCase(FixtureTestCase):
         response = self.client.get(reverse("show", kwargs=kwargs))
         self.assertEqual(response.status_code, 404)
 
-        # Different casing should redirect, MySQL always does case insensitive lookups
+        # Different casing should redirect
         kwargs["path"][-1] = self.component.slug.upper()
-        if using_postgresql():
-            response = self.client.get(reverse("show", kwargs=kwargs))
-            self.assertRedirects(
-                response,
-                self.component.get_absolute_url(),
-                status_code=301,
-            )
+        response = self.client.get(reverse("show", kwargs=kwargs))
+        self.assertRedirects(
+            response,
+            self.component.get_absolute_url(),
+            status_code=301,
+        )
 
     def test_translation_redirect(self) -> None:
         # Non existing fails with 404